refactor: split di-injection.test.js to reduce file length below 300 lines
Extract workspace-layout and tab-lifecycle DI tests into
di-injection-workspace.test.js (124 lines). Original file reduced
from 321 to 204 lines.

Closes #83
